+++ src/main/java/com/logdb/controller/DataxreceiverController.java
@@ -2,10 +2,13 @@ package com.logdb.controller;
 
 import com.logdb.dto.ClientDto;
 import com.logdb.dto.DataxceiverDto;
+import com.logdb.entity.Event;
 import com.logdb.service.ClientService;
 import com.logdb.service.DataxreceiverService;
+import com.logdb.service.EventLogService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -19,21 +22,25 @@ import java.util.Objects;
 @Controller
 public class DataxreceiverController {
 
-    protected static final String DATAXCEIVER_FORM = "/dataxceiverForm";
+    protected static final String DATAXCEIVER_FORM = "/dataxreceiverlogForm";
     protected static final String ERROR = "/error";
-    protected static final String CREATE_DATAXCEIVER = "/create/dataxceiver";
-    protected static final String DATAXCEIVER = "dataxceiver";
-    protected static final String REDIRECT_DATAXCEIVERS = "redirect:/dataxceivers/";
-    protected static final String EDIT_DATAXCEIVER_ID = "/edit/dataxceiver/{id}";
-    protected static final String DATAXCEIVER_ID = "/dataxceiver/{id}";
+    protected static final String CREATE_DATAXCEIVER = "/create/dataxreceiverlog";
+    protected static final String DATAXCEIVER = "object";
+    protected static final String REDIRECT_DATAXCEIVERS = "redirect:/dataxreceiverlog/";
+    protected static final String EDIT_DATAXCEIVER_ID = "/edit/dataxreceiverlog/{id}";
+    protected static final String DATAXCEIVER_ID = "/dataxreceiverlog/{id}";
     protected static final String REDIRECT_HOME = "redirect:/home";
-    protected static final String DATAXCEIVER_PAGE = "/dataxceiver";
+    protected static final String DATAXCEIVER_PAGE = "/dataxreceiverlogInfo";
 
     @Autowired
     protected DataxreceiverService dataxceiverService;
 
     @Autowired
     protected ClientService clientService;
+
+    @Autowired
+    EventLogService eventLogService;
+
 
 
     @RequestMapping(value = CREATE_DATAXCEIVER, method = RequestMethod.GET)
@@ -48,13 +55,15 @@ public class DataxreceiverController {
     }
 
     @RequestMapping(value = CREATE_DATAXCEIVER, method = RequestMethod.POST)
+    @Transactional
     public String createDataxceiverPost(Principal principal, @Valid DataxceiverDto dataxceiverDto, BindingResult bindingResult) {
         if (bindingResult.hasErrors()) {
             return DATAXCEIVER_FORM;
         }
         ClientDto clientDto = clientService.findAllByEmail(principal.getName());
         if (!Objects.isNull(clientDto)) {
-            dataxceiverService.insert(dataxceiverDto);
+            dataxceiverDto = dataxceiverService.insert(dataxceiverDto);
+            eventLogService.save(new Event("Created/updated dataxreceiverlog .ID:" + dataxceiverDto.getId(), clientDto.getId()));
             return REDIRECT_DATAXCEIVERS + dataxceiverDto.getId();
         }
         return ERROR;
@@ -81,11 +90,16 @@ public class DataxreceiverController {
     }
 
     @RequestMapping(value = DATAXCEIVER_ID, method = RequestMethod.DELETE)
+    @Transactional
     public String deleteDataxceiverById(@PathVariable Long id, Principal principal) {
-        DataxceiverDto dataxceiverDto = dataxceiverService.findById(id);
-        if (!Objects.isNull(dataxceiverDto)) {
+        ClientDto clientDto = clientService.findAllByEmail(principal.getName());
+        if (!Objects.isNull(clientDto)) {
+            DataxceiverDto dataxceiverDto = dataxceiverService.findById(id);
+            if (!Objects.isNull(dataxceiverDto)) {
                 dataxceiverService.delete(dataxceiverDto);
+                eventLogService.save(new Event("Deleted dataxreceiverlog .ID:" + dataxceiverDto.getId(), clientDto.getId()));
                 return REDIRECT_HOME;
+            }
         }
         return ERROR;
     }
